Move setup/cleanup logic to a separate place
